pyproject.toml
[tool.poetry]
name = "simple-telegram-bot"
version = "0.1.0"
description = ""
authors = ["teamofroman <rsateam@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
